README.md
This project's goal is to find interesting and informative aspects of how military surplus is distributed in the 1033 Program. This will especially focus on the large and expensive pieces of hardware like armored vehicles, rifles, and aircraft. Is there a reason Oklahoma gets so much of the equipment? Does is go to the southern border in greater numbers? Does it follow violent crime or officer fatalities? Or is it that certain states are more committed and better at using the program than others? These are the sorts of questions I'm looking at answering, plus maybe looking at potential outcomes.